refactor(core): publish the scenario surface from ./scenario, not ./workflow
Scenarios extend workflows; workflows should not know about scenarios. The
production workflow barrel published thirty-odd scenario and feature names,
which put a scenario module in the import graph of every server that imports
a workflow — the thing the persona-entry-point decision already forbids for
persona values.

Feature and scenario types move out of workflow.types.ts into their own
scenario.types.ts, and every scenario name is published from ./scenario.
HttpPersonasConfig and ScenarioHttpResponse lose their second home on
./services.
