+++ e2e/packages/functions/tests/scenarios/credential-oauth.steps.ts
@@ -12,7 +12,7 @@
  */
 import { pikkuScenarioStep } from '#pikku/workflow/pikku-workflow-types.gen.js'
 import { createAuthClient } from 'better-auth/client'
-import { createCookieJar, requireScenarioEnv } from '@pikku/core/workflow'
+import { createCookieJar, requireScenarioEnv } from '@pikku/core/scenario'
 
 export interface LinkUser {
   name: string
