+++ packages/cli/build.sh
@@ -157,6 +157,7 @@ while IFS= read -r -d '' f; do
       -e 's|wireSecret|defineSecret|g' \
       -e 's|wireVariable|defineVariable|g' \
       -e 's|wireCredential|defineCredential|g' \
+      -e "s|@pikku/core/internal|@pikku/core/ecosystem|g" \
       "$f" > "$tmp" && mv "$tmp" "$f"
 done < <(find .pikku \( -name '*.ts' -o -name '*.json' \) -print0)
 "$_bootstrap_dir/node_modules/.bin/pikku"
@@ -187,6 +188,7 @@ while IFS= read -r -d '' f; do
       -e "s|import { pikkuState as __pikkuState } from '@pikku/core'|import { pikkuState as __pikkuState } from '@pikku/core/ecosystem'|g" \
       -e "s|import { pikkuState as __pikkuState, CreateWireServices } from '@pikku/core'|import { pikkuState as __pikkuState, CreateWireServices } from '@pikku/core/ecosystem'|g" \
       -e "s|import { addPackageServiceFactories } from '@pikku/core'|import { pikkuState } from '@pikku/core/ecosystem'|g" \
+      -e "s|@pikku/core/internal|@pikku/core/ecosystem|g" \
       -e "s|addPackageServiceFactories('\([^']*\)', {|pikkuState('\1', 'package', 'factories', {|g" \
       -e 's|addMiddleware as addMiddlewareCore|addTagMiddleware as addTagMiddlewareCore|g' \
       -e 's|addMiddlewareCore(|addTagMiddlewareCore(|g' \
