+++ packages/cli/src/functions/wirings/scenarios/serialize-scenario-registration.test.ts
@@ -35,7 +35,8 @@ describe('the scenario wirings file', () => {
 
     assert.ok(
       output.includes(
-        `import { addWorkflow, addFeature } from '@pikku/core/workflow'`
+        `import { addWorkflow } from '@pikku/core/workflow'
+import { addFeature } from '@pikku/core/scenario'`
       )
     )
     assert.ok(
