+++ packages/core/src/public-surface.json
@@ -127,7 +127,6 @@
   "./workflow": [
     "DEFAULT_STEP_RETRIES",
     "PikkuWorkflowService",
-    "SCENARIO_SURFACES",
     "WorkflowApprovalForbiddenError",
     "WorkflowApprovalResolvedError",
     "WorkflowCancelledException",
@@ -137,35 +136,36 @@
     "WorkflowRunNotFoundError",
     "WorkflowStepFunctionMismatchError",
     "WorkflowSuspendedException",
-    "addFeature",
     "addWorkflow",
     "assertWorkflowRunOwner",
     "buildRunTimeline",
-    "composeStepProse",
-    "createCookieJar",
     "deriveInvocationId",
     "isRef",
     "pikkuWorkflowGraph",
-    "pollUntil",
     "reconstructFinalState",
     "reconstructStateAt",
-    "renderStepTemplate",
-    "requireActor",
-    "requireScenarioEnv",
-    "resolveFeatureScenarios",
-    "resolveScenarioSurfaces",
     "template",
     "uuidv5"
   ],
   "./scenario": [
     "PikkuScenarioService",
+    "SCENARIO_SURFACES",
     "ScenarioBrowserActorRequired",
     "ScenarioBrowserUnavailable",
     "ScenarioHookError",
     "ScenarioNoSurfaceBinding",
     "ScenarioNoWitness",
     "ScenarioWitnessDisagreement",
-    "createScenarioRunner"
+    "addFeature",
+    "composeStepProse",
+    "createCookieJar",
+    "createScenarioRunner",
+    "pollUntil",
+    "renderStepTemplate",
+    "requireActor",
+    "requireScenarioEnv",
+    "resolveFeatureScenarios",
+    "resolveScenarioSurfaces"
   ],
   "./workflow/timeline": [
     "buildRunTimeline",
